refactor(core): 🔨 split additional DNS verifiers
